+++ lib/styler.ex
@@ -21,10 +21,10 @@ defmodule Styler do
   @styles [
     Styler.Style.ModuleDirectives,
     Styler.Style.Pipes,
+    Styler.Style.Deprecations,
     Styler.Style.SingleNode,
     Styler.Style.Defs,
     Styler.Style.Blocks,
-    Styler.Style.Deprecations,
     Styler.Style.Configs,
     Styler.Style.CommentDirectives
   ]